xs:duration values are not ordered
Fixes #180
